+++ client-side/src/app/app.module.ts
@@ -3,7 +3,6 @@ import { HttpClientModule } from '@angular/common/http';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { AddonService } from '../services/addon.service';
-import { DataVisualizationService } from '../services/data-visualization.service';
 import { PepAddonService } from '@pepperi-addons/ngx-lib';
 import { TranslateModule, TranslateLoader, TranslateStore } from '@ngx-translate/core';
 import { AppRoutingModule } from './app.routes';
@@ -35,7 +34,6 @@ import { PepGenericListService } from '@pepperi-addons/ngx-composite-lib/generic
     providers: [
         TranslateStore,
         AddonService,
-        DataVisualizationService,
         PepGenericListService
         // When loading this module from route we need to add this here (because only this module is loading).
     ],
